+++ src/schema/user.py
@@ -8,3 +8,10 @@ class User(BaseModel):
     email: str = Field(...)
     password: str = Field(None)
     activated: bool = Field(False, description="user is activated or not")
+
+class Token(BaseModel):
+    access_token: str
+    token_type: str
+
+class TokenData(BaseModel):
+    username: Optional[str] = None
